+++ src/app/admin-module/models/data-collection-source.interface.ts
@@ -2,5 +2,5 @@ export interface DataCollectionSource {
   id: number;
   url: string;
   origin: string;
-  config: any;
+  config: Object; //
 }